bot.py
import time
from alpaca.trading.client import TradingClient
from alpaca.trading.models import Order

class AlpacaAPI:
    def __init__(self):
        self.trading_client = TradingClient(api_key="YOUR_API_KEY", secret_key="YOUR_SECRET_KEY")

    def get_account(self):
        return self.trading_client.get_account()

    def get_positions(self):
        return self.trading_client.get_positions()

    def get_penny_stocks(self, max_price, min_avg_volume=500000):
        # Implement the logic to fetch penny stocks based on the max_price and min_avg_volume
        # This is a placeholder and should be replaced with the actual implementation
        return []

    def get_historical_data(self, symbol, timeframe, limit):
        # Implement the logic to fetch historical data
        # This is a placeholder and should be replaced with the actual implementation
        return pd.DataFrame()

    def submit_order(self, symbol, qty, side, type, stop_loss, take_profit):
        # Implement the logic to submit an order
        # This is a placeholder and should be replaced with the actual implementation
        pass

def check_daily_loss(account, start_equity):
    # Implement the logic to check if the daily loss cap is reached
    # This is a placeholder and should be replaced with the actual implementation
    return True

def check_max_positions(open_positions):
    # Implement the logic to check if the max concurrent positions are reached
    # This is a placeholder and should be replaced with the actual implementation
    return True

def calculate_position_size(equity, atr, price):
    # Implement the logic to calculate position size
    # This is a placeholder and should be replaced with the actual implementation
    return 0.0

def calculate_stop_loss(price, atr):
    # Implement the logic to calculate stop loss
    # This is a placeholder and should be replaced with the actual implementation
    return 0.0

def calculate_take_profit(price, atr):
    # Implement the logic to calculate take profit
    # This is a placeholder and should be replaced with the actual implementation
    return 0.0

def generate_signals(df):
    # Implement the logic to generate trading signals
    # This is a placeholder and should be replaced with the actual implementation
    return pd.DataFrame()

def main():
    alpaca = AlpacaAPI()
    account = alpaca.get_account()
    start_equity = float(account.equity)

    while True:
        # Check daily loss cap
        account = alpaca.get_account()
        if not check_daily_loss(account, start_equity):
            print('Daily loss cap reached. Stopping trading for today.')
            break

        open_positions = alpaca.get_positions()
        if not check_max_positions(open_positions):
            print('Max concurrent positions reached. Waiting...')
            time.sleep(60)
            continue

        # Dynamically fetch penny stocks with liquidity filter
        penny_stocks = alpaca.get_penny_stocks(MAX_STOCK_PRICE, min_avg_volume=500000)
        print(f"Found {len(penny_stocks)} penny stocks to scan.")
        for symbol in penny_stocks:
            # Get latest 1-min data (last 50 bars)
            df = alpaca.get_historical_data(symbol, '1Min', limit=50)
            if df.empty or df['close'].iloc[-1] > MAX_STOCK_PRICE:
                continue

            signals = generate_signals(df)
            if signals is None:
                continue

            latest = signals.iloc[-1]
            if latest['entry_signal']:
                atr = latest['atr']
                price = latest['close']
                position_size = calculate_position_size(float(account.equity), atr, price)
                if position_size < 1:
                    continue
                stop_loss = calculate_stop_loss(price, atr)
                take_profit = calculate_take_profit(price, atr)
                print(f"Placing order: {symbol}, qty={position_size}, entry={price}, SL={stop_loss}, TP={take_profit}")
                alpaca.submit_order(
                    symbol=symbol,
                    qty=position_size,
                    side='buy',
                    type='market',
                    stop_loss=stop_loss,
                    take_profit=take_profit
                )
                time.sleep(2)  # Avoid API rate limits

        # Sleep before next scan
        time.sleep(60)

if __name__ == '__main__':
    main() 