requirements.txt
alpaca-trade-api
pandas
numpy
ta
python-dotenv
websockets
Flask 